LayoutTests/inspector/unit-tests/target-manager-expected.txt

Test for TargetManager and other global WI.Target objects.


== Running test suite: TargetManager
-- Running test case: TargetManager.MainTarget
PASS: Without site isolation, targets list should always start out with the main page target only.
PASS: Target list should always contain the main page target.
PASS: Main target should have an ExecutionContext.
PASS: Main target should have the page target's RuntimeAgent.
Target - page - Page

-- Running test case: TargetManager.WorkerTarget.Create
PASS: Added Target should have Worker type.
PASS: Added Target should have an ExecutionContext.
PASS: Added Target should have a RuntimeAgent.
PASS: Added Target RuntimeAgent should not be the global RuntimeAgent.
Target - page - Page
Target - worker - worker-1.js

-- Running test case: TargetManager.WorkerTarget.Remove
PASS: Removed Target should have Worker type.
Target - page - Page

-- Running test case: TargetManager.FrameTarget.Create
Target - page - Page

-- Running test case: TargetManager.FrameTarget.Remove
Target - page - Page
